\name{euspca-package}
\alias{euspca-package}
\docType{package}
\title{
Exactly uncorrelated sparse principal component analysis
}
\description{
Exactly uncorrelated sparse principal component analysis (EUSPCA) is an algorithm to find uncorrelated sparse principal components. This package solves EUSPCA with \eqn{\ell_1}{l1} regularization. The algorithm is based on a combination of augmented Lagrangian method and non-monotone proximal gradient method. \cr
}

\details{
The main function of the package \code{\link{euspca}} is \code{euspca}. \cr
}

\keyword{euspca}
\keyword{principal component analysis}
\keyword{sparse principal component}
\keyword{uncorrelated component}
\keyword{package}

