\name{print.euspca}
\alias{print.euspca}
\title{Print method for euspca objects}
\description{
Print out a euspca fit. 
}
\usage{
\method{print}{euspca}(x, print.loadings=FALSE, ...)  
}
\arguments{
  \item{x}{euspca object}
  \item{print.loadings}{Print the loadings or not? Default is \code{FALSE}.}
\item{\dots}{Additonal arguments for generic print.}
}
\value{NULL}
\keyword{methods}
